# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoices",
        "version": "1.0"
    },
    "paths": {
        "/api/invoices": {
            "get": {
                "operationId": "2e947f49543378e8d1ac24fc3ebcdcc7",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "operationId": "15892ddc0ad0649e0919d15345232fc5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/invoices/{id}": {
            "get": {
                "operationId": "5d25d392f44d2cf516114d871e72ecb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do registro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "operationId": "e5aa8abff3efb646c72e59724ecb13e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do registro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "operationId": "410b8a2e4b097016c3dea391349a789b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do registro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    }
}